correcting typos and making the doc more concise
I was going through the documentation and couldn't help noticing a few typos, and some lack of celerity (Especially to English readers).
It's really not a big deal and some people will just put the dots together. But clarity might help some newbies feel it's less technical.
Finally I added "Azure" to the list of cloud providers, because I feel no list of cloud providers is complete without the second largest cloud providers in the world.